refactor: 테스트 케이스 생성 함수 리팩터링
