+++ client/package.json
@@ -7,7 +7,8 @@
     "dev": "vite",
     "build": "tsc && vite build",
     "preview": "vite preview",
-    "test": "jest --coverage && codecov"
+    "test": "jest --coverage && codecov",
+    "test:local" : "jest"
   },
   "dependencies": {
     "react": "^18.2.0",
